+++ CHANGELOG.md
@@ -10,6 +10,8 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 
+- Retained edits that cannot be received can now be removed from this device after confirmation, clearing their warning without changing current documents.
+
 ## [0.3.0] - 2026-09-20
 
 ### Added